Create LightMonster class which will hold a reference to the Monster class which has all the stat info. LightMonster will store EVs and the current hp, but only have getters for the other stats. SInce the game is turn based, there's no need to store the values for high speed access. Can calculate values on demand.

+++ Assets/Scripts/Monster.cs
@@ -30,7 +30,7 @@ public class Monster : MonoBehaviour
 
     private void InitializeBaseTotal()
     {
-        BaseStatTotal = (short)0;
+        BaseStatTotal = 0;
         BaseStatTotal += BaseHP;
         BaseStatTotal += BaseAttack;
         BaseStatTotal += BaseDefense;
